fixup: Move useExecutor to a mutations directory
